23680961: get film test
